+++ libs/rxjs/utility/src/lib/from-readable-stream.ts
@@ -2,12 +2,12 @@
  * @packageDocumentation
  * @module Utility
  */
-import { Observable } from 'rxjs';
+import { Observable, Subscriber } from 'rxjs';
 
 /**
  * Creates an Observable source from a
- * {@link https://developer.mozilla.org/en-US/docs/Web/API/ReadableStream|ReadableStream} source that will emit any values
- * emitted by the stream.
+ * {@link https://developer.mozilla.org/en-US/docs/Web/API/ReadableStream|ReadableStream} source that will emit any
+ * values emitted by the stream.
  *
  * @category Streams
  *
@@ -15,7 +15,7 @@ import { Observable } from 'rxjs';
  * @param signal Optional {@link https://developer.mozilla.org/en-US/docs/Web/API/AbortSignal|AbortSignal} to provide
  *   to the underlying stream
  * @param queueStrategy Optional strategy for backpressure queueing
- * @param throwAbortAsError Optional boolean to throw an `AbortController` signal as an error instead of just completing
+ * @param throwEndAsError Optional to return an error when the `AbortSignal` has been fired instead of just closing
  *
  * @example Create a ReadableStream of `0` to `100` and convert to an Observable
  * ```ts
@@ -38,31 +38,37 @@ export function fromReadableStream<T extends unknown>(
   stream: ReadableStream<T>,
   signal?: AbortSignal,
   queueStrategy?: QueuingStrategy,
-  throwAbortAsError = false,
+  throwEndAsError = false,
 ): Observable<T> {
+  /**
+   * @private
+   * @internal
+   * @param subscriber
+   */
+  function createStream(subscriber: Subscriber<T>) {
+    return new WritableStream<T>(
+      {
+        write: (value) => subscriber.next(value),
+        abort: (error) => {
+          if (throwEndAsError) {
+            subscriber.error(error);
+          } else if (!subscriber.closed) {
+            subscriber.complete();
+          }
+        },
+        close: () => {
+          if (!subscriber.closed) {
+            subscriber.complete();
+          }
+        },
+      },
+      queueStrategy,
+    );
+  }
+
   return new Observable<T>((subscriber) => {
     stream
-      .pipeTo(
-        new WritableStream<T>(
-          {
-            write: (value) => subscriber.next(value),
-            abort: (error) => {
-              if (throwAbortAsError) {
-                subscriber.error(error);
-              } else if (!subscriber.closed) {
-                subscriber.complete();
-              }
-            },
-            close: () => {
-              if (!subscriber.closed) {
-                subscriber.complete();
-              }
-            },
-          },
-          queueStrategy,
-        ),
-        { signal },
-      )
+      .pipeTo(createStream(subscriber), { signal })
       .then(() => !subscriber.closed && subscriber.complete())
       .catch((error) => subscriber.error(error));