+++ libs/rxjs/utility/src/lib/to-writable-stream.spec.ts
@@ -1,6 +1,6 @@
 import { observe } from 'rxjs-marbles/jest';
 import { toWritableStream } from '@rxjs-ninja/rxjs-utility';
-import { catchError, reduce, switchMap, take, tap } from 'rxjs/operators';
+import { catchError, map, reduce, take, tap } from 'rxjs/operators';
 import { from, of, throwError } from 'rxjs';
 import { WritableStream } from 'web-streams-polyfill/ponyfill';
 
@@ -47,6 +47,53 @@ describe('toWritableStream', () => {
         close: () => expect(output).toBe('123'),
       });
       return from([1, 2, 3, 4, 5]).pipe(take(3), toWritableStream(stream));
+    }),
+  );
+
+  it(
+    'should end writing to the writable stream on error',
+    observe(() => {
+      let output = '';
+      const stream = new WritableStream({
+        write: (val) => {
+          output += val;
+        },
+        close: () => expect(output).toBe('123'),
+      });
+
+      return from([1, 2, 3, 4, 5]).pipe(
+        map((val) => {
+          if (val < 4) {
+            return val;
+          } else {
+            return throwError('Number is 4');
+          }
+        }),
+        toWritableStream(stream),
+        catchError(() => of(true)),
+      );
+    }),
+  );
+
+  it(
+    'should no longer write when stream is closed',
+    observe(() => {
+      const ctrl = new AbortController();
+
+      let output = '';
+      const stream = new WritableStream({
+        write: (val) => {
+          output += val;
+        },
+        close: () => expect(output).toBe('123'),
+      });
+
+      return from([1, 2, 3, 4, 5]).pipe(
+        tap((val) => val === 4 && ctrl.abort()),
+        toWritableStream(stream, ctrl.signal),
+        reduce((a, b) => a + b, 0),
+        tap((val) => expect(val).toBe(15)),
+      );
     }),
   );
 });